Calculate the MP3 duration correctly
Now this finally matches the duration produced by `sox file.mp3 -n stat`
and `ffmpeg -v quiet -stats -i file.mp3 -f null -` (see [SO][] for the
source of the commands)!  Although not exactly because the parser does
include the information frames in the calculation. The issue was with
the slightly incorrect duration of one MP3 frame.

What's more surprising is that the compiled program calculates the
duration for a 193 MiB file (a recording for 04:40:32) faster (7.1 s)
than `sox` (19.3 s) and `ffmpeg` (13.4 s)! This needs more
investigation. Also, replacing `megaparsec` with `attoparsec` could make
parsing even faster (according to `megaparsec`'s README).

[SO]: https://stackoverflow.com/questions/10437750/how-to-get-the-real-actual-duration-of-an-mp3-file-vbr-or-cbr-server-side