Test that `frameParser` parses a frame
In fact, the parser always returns a success because there is no test to
contradict that yet.

`attoparsec` is used (instead of `megaparsec`) for the new parser
because it should be faster (which is more important on bigger files)
and already works with `conduit` (this feature is not used now).