Show Cyrillic filepaths correctly
It still puzzles (and annoys) me to no end that the default `Show`
instance for strings and text prints non-ASCII characters octal-escaped,
and the only way to print them properly is to use `putStrLn`-like
functions. However I don't print them myself here, `it` does that, so I
have to use the `ushow` hack.