Fix the generation of `FramesWithMiddleJunk`
I'm not sure how it worked before. Looks like `framesBefore` and
`framesAfter` were of type `[ByteString]`, which are invalid frames, so
the test would always succeed (failed parsing) even if `junk` is empty
(since it uses `Arbitrary ByteString`, which can generate empty data).

Now the types are explicitly `ValidMP3Frame`s, so the test is fixed.