Decrease the verified duration error to 0.1 seconds
The test still passes with 588/588 (100%) successful examples.

This error should be enough for my purposes; decreasing it to 0.01
seconds causes failures. They must be because the parser doesn't skip
the Xing/Info/VBRI/Lame header frame, which is typically 0.026 s long.
If there are 200 new podcast episodes, the max error will be 20 seconds;
assuming a (very) average duration of 45 minutes/episode, the total time
is 150 hours, of which the 20 seconds is mere 0.004%! Good.