Explicitly calculate duration only if there is no cached value
This avoids the calculations even though there are cached durations in
the program; the unit test didn't uncover this issue. Apparently the
reason is in the different, stricter (?) monad `CachedDurationM`.

I've failed to replicate this bug with the unit test by:

* moving the read-only maps from the `Reader` into the `State` monad;

* making a strict data type (effectively, a tuple) with the read-only
  maps (the mock environment) for the `Reader`;

* replacing the strict maps with lazy ones;

* replacing the duration map with a function.

None of these approaches could distinguish the difference in the
implementation: they either succeeded, or failed in both cases.

¯\_(ツ)_/¯